+++ helper/ArrayStorageFormatter.php
@@ -1,0 +1,10 @@
+<?php
+namespace sergeylisitsyn\settingsStorage\helper;
+
+class ArrayStorageFormatter
+{
+    public function format($value)
+    {
+        return unserialize($value);
+    }
+}