+++ helper/BooleanStorageFormatter.php
@@ -1,0 +1,10 @@
+<?php
+namespace sergeylisitsyn\settingsStorage\helper;
+
+class BooleanStorageFormatter
+{
+    public function format($value)
+    {
+        return (int) $set->value ? true : false;
+    }
+}