+++ helper/NumberStorageFormatter.php
@@ -1,0 +1,10 @@
+<?php
+namespace sergeylisitsyn\settingsStorage\helper;
+
+class NumberStorageFormatter
+{
+    public function format($value)
+    {
+        return $value;
+    }
+}