+++ helper/StringStorageFormatter.php
@@ -1,0 +1,11 @@
+<?php
+namespace sergeylisitsyn\settingsStorage\helper;
+
+class StringStorageFormatter
+{
+    
+    public function format($value)
+    {
+        return $value;
+    }
+}